+++ packages/f-vue-icons/src/components/RestaurantIcon/RestaurantIcon.vue
@@ -1,0 +1,16 @@
+<template>
+    <restaurant-icon :class="['c-icon c-icon--restaurant', iconClasses]" />
+</template>
+
+<script>
+import RestaurantIcon from '@justeat/f-icons/src/img/icons/general/restaurant.svg';
+import iconsMixin from '../../mixins/icons.mixin';
+
+export default {
+    components: {
+        RestaurantIcon
+    },
+
+    mixins: [iconsMixin]
+};
+</script>