+++ packages/f-vue-icons/src/components/RestaurantIcon/index.js
@@ -1,0 +1,7 @@
+import RestaurantIcon from './RestaurantIcon.vue';
+
+const install = Vue => {
+    Vue.component(RestaurantIcon.name, RestaurantIcon);
+};
+
+export default install;