+++ packages/f-vue-icons/src/components/index.js
@@ -19,6 +19,8 @@ import MinusIcon from './MinusIcon';
 import OfferIcon from './OfferIcon';
 import PadlockIcon from './PadlockIcon';
 import PlusIcon from './PlusIcon';
+import RestaurantIcon from './RestaurantIcon';
+import SortIcon from './SortIcon';
 import SpicyIcon from './SpicyIcon';
 import StopwatchIcon from './StopwatchIcon';
 import TickIcon from './TickIcon';
@@ -47,6 +49,8 @@ export {
     OfferIcon,
     PadlockIcon,
     PlusIcon,
+    RestaurantIcon,
+    SortIcon,
     SpicyIcon,
     StopwatchIcon,
     TickIcon,