+++ packages/f-form-field/CHANGELOG.md
@@ -4,6 +4,28 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.5.0
+------------------------------
+*May 29, 2020*
+
+### Added
+- `data-test-id` attribute to the input element.
+
+
+v0.4.0
+------------------------------
+*May 26, 2020*
+
+### Added
+- Component now emits an event when its value is updated. This is so that validation applied to the consuming form component knows when to run on the field.
+- Unique ID applied to form input so that clashes between fields doesn't occur.
+- Named slot for error message markup to be passed through from a consuming form.
+
+### Changed
+- Wrapper element added around the input field, to accommodate error message placement.
+- `for` attribute moved from inside the `FormLabel` component and is now applied in the `FormField` component, so that it can make use of passed in `$attrs`.
+
+
 v0.3.0
 ------------------------------
 *May 18, 2020*